requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
openai>=1.12.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pandas>=2.1.0
